requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
websockets>=12.0
python-dotenv>=1.0.0
